Fix test collection: rename orchestrator/auth.py to lifecycle_auth.py
The bare `from auth import require_lifecycle_secret` resolved to
gateway/auth.py in CI because gateway/tests/conftest.py registers
gateway's auth module as sys.modules["auth"] during collection
(gateway/tests comes before orchestrator/tests in testpaths).
Renaming to lifecycle_auth.py eliminates the module-name collision.
