+++ orchestrator/routes/phases.py
@@ -31,7 +31,7 @@ except ImportError:
         return logging.getLogger(name)
 
 
-from auth import require_lifecycle_secret
+from lifecycle_auth import require_lifecycle_secret
 from models import (
     PipelinePhase,
     PipelineStatus,
