+++ orchestrator/tests/test_decisions_routes.py
@@ -1075,7 +1075,7 @@ class TestLifecycleSecretAuth:
         its audit log line. We verify the attach directly because egg's
         structlog output bypasses pytest's caplog fixture.
         """
-        from auth import require_lifecycle_secret
+        from lifecycle_auth import require_lifecycle_secret
         from flask import Flask, request
 
         seen: dict[str, str] = {}
